fix: facebook logout button style
